requirements.txt
torch
torchvision
lightning
timm
pandas
numpy
scikit-learn
Pillow
matplotlib
transformers


